combine authInfo and auth
